Testing scan parsing for various os
